modify: BLOCK_DATA 파일 생성 및 import
